+++ test/src/analyzers/lint_analyzer/reporters/reporters_list/console/lint_console_reporter_test.dart
@@ -52,13 +52,13 @@ void main() {
         equals(
           [
             'test/resources/abstract_class.dart:',
-            '\x1B[38;5;11mWarning \x1B[0mmetric1: \x1B[38;5;11m100\x1B[0m',
+            '\x1B[38;5;11mWarning \x1B[0mmetric1: \x1B[38;5;11m100 units\x1B[0m',
             '\x1B[38;5;9mAlarm   \x1B[0mclass.constructor - metric2: \x1B[38;5;9m10\x1B[0m',
             '',
             'test/resources/class_with_factory_constructors.dart:',
             '\x1B[38;5;11mWarning \x1B[0msimple message : 0:0 : id',
             '\x1B[38;5;4mStyle   \x1B[0msimple design message : 0:0 : designId',
-            '\x1B[38;5;11mWarning \x1B[0mfunction - metric4: \x1B[38;5;11m5\x1B[0m',
+            '\x1B[38;5;11mWarning \x1B[0mfunction - metric4: \x1B[38;5;11m5 units\x1B[0m',
             '',
           ],
         ),
@@ -68,7 +68,7 @@ void main() {
         equals(
           [
             'test/resources/abstract_class.dart:',
-            '\x1B[38;5;11mWarning \x1B[0mmetric1: \x1B[38;5;11m100\x1B[0m',
+            '\x1B[38;5;11mWarning \x1B[0mmetric1: \x1B[38;5;11m100 units\x1B[0m',
             '\x1B[38;5;7m        \x1B[0mclass - metric1: \x1B[38;5;7m0\x1B[0m',
             '\x1B[38;5;9mAlarm   \x1B[0mclass.constructor - metric2: \x1B[38;5;9m10\x1B[0m',
             '\x1B[38;5;7m        \x1B[0mclass.method - metric3: \x1B[38;5;7m1\x1B[0m',
@@ -77,7 +77,7 @@ void main() {
             '\x1B[38;5;7m        \x1B[0mmetric1: \x1B[38;5;7m0\x1B[0m, metric2: \x1B[38;5;7m1\x1B[0m',
             '\x1B[38;5;11mWarning \x1B[0msimple message : 0:0 : id',
             '\x1B[38;5;4mStyle   \x1B[0msimple design message : 0:0 : designId',
-            '\x1B[38;5;11mWarning \x1B[0mfunction - metric4: \x1B[38;5;11m5\x1B[0m',
+            '\x1B[38;5;11mWarning \x1B[0mfunction - metric4: \x1B[38;5;11m5 units\x1B[0m',
             '',
           ],
         ),