+++ test/src/analyzers/lint_analyzer/reporters/reporters_list/json/lint_json_reporter_test.dart
@@ -44,6 +44,7 @@ void main() {
           {
             'metricsId': 'file-metric-id',
             'value': 100,
+            'unitType': 'units',
             'level': 'warning',
             'comment': 'metric comment',
             'context': <String>[],